requirements.txt
Appium-Python-Client==2.9.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
pip==23.0.1
PySocks==1.7.1
selenium==4.8.3
setuptools==67.6.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio-websocket==0.10.2
trio==0.22.0
urllib3==1.26.15
wheel==0.40.0
wsproto==1.2.0
